+++ anyblok/bloks/anyblok_core/system/sequence.py
@@ -33,6 +33,14 @@ class Sequence:
         seq = SQLASequence(cls._cls_seq_name)
         seq.create(cls.registry.bind)
 
+        if hasattr(cls.registry, '_wanted_sequence'):
+            if cls.registry._wanted_sequence:
+                for sequence in cls.registry._wanted_sequence:
+                    if sequence['formater'] is None:
+                        del sequence['formater']
+
+                    cls.insert(**sequence)
+
     @classmethod
     def create_sequence(cls, values):
         """ create the sequence for one instance """