+++ doc/CHANGES.rst
@@ -62,6 +62,7 @@ Future
 * [IMP] in column the default can be a classmethod name
 * [REF] replace all the field (prefix, suffic, ...) by a formater field.
   It is a python formater string
+* [IMP] Sequence column
 
 0.4.1
 -----